break action menu into owner, editor, upload
